requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
connectorx==0.3.1
decorator==5.1.1
diot==0.2.2
funda-scraper==1.0.8
idna==3.4
inflection==0.5.1
lxml==4.9.3
numpy==1.25.2
pandas==2.0.3
polars==0.18.15
py==1.11.0
pyarrow==13.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
retry==0.9.2
six==1.16.0
soupsieve==2.4.1
tqdm==4.66.1
tzdata==2023.3
urllib3==1.26.0
